Create Plugin first scratch with another ponit of view
Implementing the core functionality of the stock list feature.
With this point of view we are saving a clone of the products in
the same table of Magento, and keeping a reference.

We have also a test admin to set the prodcts to a franchise before
coding the observer at the order event.